reference/experimental/tritrpc/pyproject.toml
[project]
name = "tritrpc-requirements-impl"
version = "0.4.0"
description = "Reference implementation for TriTRPC vNext framing, compliance gates, and braided identity."
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "jsonschema>=4.0", "cryptography>=42.0", "protobuf>=5.0", "thrift>=0.20.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
